refactor(app): Separate all application scenarios from `StateActor` into `Application`
