+++ app/src/dependency.rs
@@ -1,12 +1,20 @@
+use std::collections::HashMap;
 #[cfg(any(feature = "test-doubles", test))]
 pub use test_doubles::TestDependencies;
 
 use {
-    move_core_types::effects::ChangeSet, moved_blockchain::payload::PayloadId,
-    moved_genesis::config::GenesisConfig, moved_shared::primitives::B256,
+    move_core_types::effects::ChangeSet,
+    moved_blockchain::payload::PayloadId,
+    moved_execution::{L1GasFeeInput, transaction::ExtendedTxEnvelope},
+    moved_genesis::config::GenesisConfig,
+    moved_shared::primitives::B256,
 };
 
 pub struct Application<D: Dependencies> {
+    pub genesis_config: GenesisConfig,
+    pub head: B256,
+    pub height: u64,
+    pub mem_pool: HashMap<B256, (ExtendedTxEnvelope, L1GasFeeInput)>,
     pub gas_fee: D::BaseGasFee,
     pub base_token: D::BaseTokenAccounts,
     pub l1_fee: D::CreateL1GasFee,
@@ -32,6 +40,10 @@ pub struct Application<D: Dependencies> {
 impl<D: Dependencies> Application<D> {
     pub fn new(_: D, genesis_config: &GenesisConfig) -> Self {
         Self {
+            genesis_config: genesis_config.clone(),
+            head: Default::default(),
+            height: 0,
+            mem_pool: Default::default(),
             gas_fee: D::base_gas_fee(),
             base_token: D::base_token_accounts(genesis_config),
             l1_fee: D::create_l1_gas_fee(),
