+++ app/src/lib.rs
@@ -3,4 +3,9 @@ pub use {actor::*, dependency::*, input::*};
 pub(crate) mod input;
 
 mod actor;
+mod command;
 mod dependency;
+mod query;
+
+#[cfg(test)]
+mod tests;
